+++ CHANGELOG.md
@@ -8,8 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 * Set up project and CI/CD GitHub actions, [PR-9](https://github.com/reductstore/reduct-go/pull/9)
 * Implement Bucket API [PR-10](https://github.com/reductstore/reduct-go/pull/10)
 * Implement Entry API [PR-12](https://github.com/reductstore/reduct-go/pull/12)
 * Implement Server API [PR-14](https://github.com/reductstore/reduct-go/pull/14)
 * Implement Token API [PR-15](https://github.com/reductstore/reduct-go/pull/15)
+* Implement Replication API [PR-16](https://github.com/reductstore/reduct-go/pull/16)
+* Improvement for Bucket API [PR-18](https://github.com/reductstore/reduct-go/pull/18)